refactor(#45): rebase and fix notification routes
